Load existing breakpoints from the cell
